+++ SurvivalGame/Assets/WorldObjects/Building.cs
@@ -11,6 +11,7 @@ public interface IBuildable
 {
     void Build(GameManager gm);
     bool CanBuild(GameManager gm);
+    bool InteractWithBuilding();
 }
 
 [DisallowMultipleComponent]